add tests for retry policies
